+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- Fixed a bug where the `SchwiftyBuilder` widget would create a new instance of `Schwifty` instead of using the provided instance.
+- Added `example/main.dart` to demonstrate the usage of `schwifty`.
+- Updated the README.
+
 ## 0.0.1
 
 - Initial release of `schwifty`.